requirements.txt
Django==2.2
djangorestframework==3.10.3
django-filter==2.2.0
sqlparse
psycopg2
django-extensions
